# Fix accuracy of max-pooling backpropagation for bfloat16 data