… src/cpu/x64/jit_uni_pooling.hpp
… tests/benchdnn/inputs/pool/test_pool_bfloat16
… tests/benchdnn/pool/pool.cpp