@asimonov1 2025-01-14
make test
disable device_gpu
disable benchdnn_all
enable benchdnn_pool
enable benchdnn_nightly